add basic source option for loading template files
This class models the root that template files will be loaded from
and implements how they will be rendered and compiled.  This will
be used by the engine to implement the Deas rendering API.

All this really does is wrap calls to `Kramdown::Document` and allows
optionally caching template content.  This closely follows the pattern
setup by deas-erubis.